requirements.txt
affine==2.4.0
asttokens==3.0.0
attrs==25.3.0
blessings==1.7
bqplot==0.12.45
branca==0.8.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1
cligj==0.7.2
colour==0.1.5
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
earthengine-api==1.5.19
eerepr==0.1.2
executing==2.2.0
filelock==3.18.0
folium==0.19.7
fonttools==4.58.2
fsspec==2025.5.1
future==1.0.0
geemap==0.35.3
geocoder==1.38.1
google-api-core==2.25.0
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipyleaflet==0.19.2
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipytree==0.2.2
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jupyter-leaflet==0.19.2
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==1.42.0
networkx==3.5
numpy==2.3.0
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
plotly==6.1.2
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
pyparsing==3.2.3
pyperclip==1.9.0
pyshp==2.3.1
python-box==7.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
rasterio==1.4.3
ratelim==0.1.6
requests==2.32.4
rsa==4.9.1
scooby==0.10.1
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
torch==2.7.1
traitlets==5.14.3
traittypes==0.2.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
xyzservices==2025.4.0
